extracted dropdown & checkbox logic
