+++ src/components/ColorSelectorWidget.tsx
@@ -1,4 +1,4 @@
-import React, { useRef, useState } from "react";
+import React, { useState } from "react";
 import styled from "styled-components";
 import { Swatch } from "../styles/Swatch";
 import useBreakpointMap from "../hooks/useBreakpointMap";
@@ -7,46 +7,19 @@ import WheelPicker from "./Pickers/WheelPicker";
 import ColorIndicator, { Heading } from "./ColorIndicator";
 import useDebounce from "../hooks/useDebounce";
 import useSliderChange from "../hooks/useSliderChange";
-import CM, { ColorMaster, extendPlugins } from "colormaster";
+import { ColorMaster, extendPlugins } from "colormaster";
 import NamePlugin from "colormaster/plugins/name";
-import { FlexColumn } from "./Sliders/RGBSliderGroup";
-import { FlexRow } from "./Sliders/FullSlider";
 import Spacers from "./Spacers";
 import Dropdown from "./Dropdown";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import {
-  faAngleDown,
-  faAngleLeft,
-  faAngleRight,
-  faAngleUp,
-  faCrosshairs,
-  faPalette
-} from "@fortawesome/free-solid-svg-icons";
+import { faCrosshairs, faPalette } from "@fortawesome/free-solid-svg-icons";
+import SwatchCarousel from "./SwatchCarousel";
+import { FlexColumn, FlexRow } from "../styles/Flex";
 
 extendPlugins([NamePlugin]);
 
 const colorspaceOpts = ["rgb", "hex", "hsl"];
 const pickerOpts = ["slider", "sketch", "wheel"];
-
-const SWATCH_COLORS = [
-  "hsla(0, 100%, 50%, 1)",
-  "hsla(30, 100%, 50%, 1)",
-  "hsla(60, 100%, 50%, 1)",
-  "hsla(90, 100%, 50%, 1)",
-  "hsla(120, 100%, 50%, 1)",
-  "hsla(150, 100%, 50%, 1)",
-  "hsla(180, 100%, 50%, 1)",
-  "hsla(210, 100%, 50%, 1)",
-  "hsla(240, 100%, 50%, 1)",
-  "hsla(270, 100%, 50%, 1)",
-  "hsla(300, 100%, 50%, 1)",
-  "hsla(330, 100%, 50%, 1)",
-  "hsla(0, 0%, 100%, 1)",
-  "hsla(0, 0%, 75%, 1)",
-  "hsla(0, 0%,50%, 1)",
-  "hsla(0, 0%,25%, 1)",
-  "hsla(0, 0%, 0%, 1)"
-];
 
 const BorderedSegment = styled.div`
   border: 1px solid hsla(0, 0%, 75%, 1);
@@ -55,7 +28,7 @@ const BorderedSegment = styled.div`
   box-shadow: 5px 5px 10px 1px rgba(102, 102, 102, 0.5);
 `;
 
-const StyledAngleIcon = styled(FontAwesomeIcon).attrs((props: { $disabled: boolean }) => props)`
+export const StyledAngleIcon = styled(FontAwesomeIcon).attrs((props: { $disabled: boolean }) => props)`
   cursor: ${(props) => (props.$disabled ? "not-allowed" : "pointer")};
 
   & path {
@@ -66,6 +39,7 @@ const StyledAngleIcon = styled(FontAwesomeIcon).attrs((props: { $disabled: boole
     }
   }
 `;
+
 interface IColorSelectorWidget {
   color: ColorMaster;
   setColor: React.Dispatch<React.SetStateAction<ColorMaster>>;
@@ -83,10 +57,7 @@ export default function ColorSelectorWidget({
   initPicker = pickerOpts[0],
   harmony = undefined
 }: IColorSelectorWidget): JSX.Element {
-  const numVisibleSwatches = useRef(9);
-
   const [alpha, setAlpha] = useState(true);
-  const [swatchIndex, setSwatchIndex] = useState(0);
   const [colorspace, setColorspace] = useState(colorspaceOpts.find((x) => x === initColorspace) ?? colorspaceOpts[0]);
   const [picker, setPicker] = useState(pickerOpts.find((x) => x === initPicker) ?? pickerOpts[0]);
 
@@ -94,82 +65,33 @@ export default function ColorSelectorWidget({
   const { isComputer, isWideScreen } = useBreakpointMap();
   const currentSliders = useSliderChange({ color, setColor, colorspace, alpha });
 
-  const adjustSelection = (dir: "up" | "down", type: "picker" | "colorspace") => {
-    const opts = type === "colorspace" ? colorspaceOpts : pickerOpts;
-    const value = type === "colorspace" ? colorspace : picker;
-    let newValue = "";
-    const currentIndex = opts.indexOf(value);
-    if (dir === "down") {
-      newValue = currentIndex === 0 ? opts[opts.length - 1] : opts[currentIndex - 1];
-    } else {
-      newValue = currentIndex === opts.length - 1 ? opts[0] : opts[currentIndex + 1];
-    }
-
-    if (type === "colorspace") setColorspace(newValue);
-    else setPicker(newValue);
-  };
-
   return (
     <BorderedSegment>
       {children}
 
       <FlexColumn>
-        <Swatch $radius={50} background={color.stringHSL()} title={color.stringHSL()} $cursor="help" />
+        <FlexRow>
+          <Swatch $radius={50} background={color.stringHSL()} title={color.stringHSL()} $cursor="help" />
+        </FlexRow>
+
+        <Spacers height="15px" />
 
         <Heading color="grey">{colorNameDebounce}</Heading>
 
-        <FlexRow>
-          <StyledAngleIcon
-            icon={faAngleLeft}
-            color="gray"
-            $disabled={swatchIndex === 0}
-            onClick={() => swatchIndex > 0 && setSwatchIndex(swatchIndex - 1)}
-          />
+        <Spacers height="15px" />
 
-          <Spacers width="4px" />
-
-          {SWATCH_COLORS.slice(swatchIndex, swatchIndex + numVisibleSwatches.current).map((background) => (
-            <Swatch
-              className="swatch-color"
-              key={background + "-swatch"}
-              title={background}
-              $radius={15}
-              $borderColor="rgba(0,0,0,0.3)"
-              $borderRadius="4px"
-              display="inline-block"
-              background={background}
-              onClick={() => setColor(CM(background))}
-            />
-          ))}
-
-          <Spacers width="4px" />
-
-          <StyledAngleIcon
-            icon={faAngleRight}
-            color="gray"
-            $disabled={swatchIndex === SWATCH_COLORS.length - numVisibleSwatches.current}
-            onClick={() =>
-              swatchIndex < SWATCH_COLORS.length - numVisibleSwatches.current && setSwatchIndex(swatchIndex + 1)
-            }
-          />
-        </FlexRow>
+        <SwatchCarousel setColor={setColor} />
 
         <Spacers height="30px" />
 
         <FlexRow>
-          <FlexColumn>
-            <StyledAngleIcon icon={faAngleUp} color="gray" onClick={() => adjustSelection("up", "colorspace")} />
-            <StyledAngleIcon icon={faAngleDown} color="gray" onClick={() => adjustSelection("down", "colorspace")} />
-          </FlexColumn>
-
-          <Spacers width="5px" />
-
           <Dropdown
             opts={colorspaceOpts}
             value={colorspace}
             setValue={setColorspace as React.Dispatch<React.SetStateAction<string>>}
             icon={<FontAwesomeIcon icon={faPalette} color="dimgray" />}
             iconPos="left"
+            switcherPos="left"
             cols={8}
           />
 
@@ -181,22 +103,16 @@ export default function ColorSelectorWidget({
             setValue={setPicker as React.Dispatch<React.SetStateAction<string>>}
             icon={<FontAwesomeIcon icon={faCrosshairs} color="dimgray" />}
             iconPos="right"
+            switcherPos="right"
             cols={8}
           />
-
-          <Spacers width="5px" />
-
-          <FlexColumn>
-            <StyledAngleIcon icon={faAngleUp} color="gray" onClick={() => adjustSelection("up", "picker")} />
-            <StyledAngleIcon icon={faAngleDown} color="gray" onClick={() => adjustSelection("down", "picker")} />
-          </FlexColumn>
         </FlexRow>
 
-        <Spacers height="30px" />
+        <Spacers height="20px" />
 
         <ColorIndicator color={currentSliders.colorStr} showName={false} alpha={alpha} setAlpha={setAlpha} />
 
-        <Spacers height="30px" />
+        <Spacers height="20px" />
 
         {picker === "sketch" ? (
           <SketchPicker color={color} setColor={setColor} verticalPickers={isComputer || isWideScreen} />
