+++ src/styles/Canvas.ts
