+++ src/styles/Swatch.ts
@@ -20,9 +20,6 @@ export const Swatch = styled.div.attrs(
   height: ${(props) => props.$radius * 2 + (props.$units ?? "px")};
   border-radius: ${(props) => props.$borderRadius ?? "50%"};
   border: ${(props) => props.$borderColor ?? "hsla(0, 0%, 90%, 1)"} 1px solid;
-  display: ${(props) => props.display ?? "block"};
-  margin: auto ${(props) => (props.display ? "2px" : "auto")};
-  position: ${(props) => props.position ?? ""};
   cursor: ${(props) => props.$cursor};
 `;
